Deleted Unused Commands, Reorganized others into new package
