+++ src/main/java/frc/robot/commands/autonomous/AutoMove.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.autonomous;
 
 import frc.robot.Robot;
 import edu.wpi.first.wpilibj2.command.CommandBase;
